prompt-log-server/requirements.txt
flask
sqlalchemy
psycopg2-binary
flask-cors
python-dotenv